requirements.txt
astroid==1.6.3
colorama==0.3.9
Django==2.0.5
django-tastypie==0.14.1
djangorestframework==3.8.2
flake8==3.5.0
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.4
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.4
six==1.11.0
wrapt==1.10.11
